No more fatal in Repository
Repository shouldn't call fmt.Fatal(...), instead it should fail gracefully by setting the state of a repo to StateError.

This could happen if e.g. a folder contains a .git-folder, but is'nt actually valid or corrupted. So we make a Repository but can't use it at all and it will fail immediatly.